OLLbase: Activate lilyglyphs (XeLaTeX necessary!)
There was a bug in the lilyglyphs that prevented it
to work in newer LaTeX distributions (don't know which ones).
Obviously it was a wrong syntax in the first place and I seem
to have fixed it.

Therefore I now activate the lilyglyphs package that
had so far been commented out.

!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
This means that now all OLL LaTeX files have to be compiled
with XeLaTeX due to the use of fontspec.
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!