LaTeX files: new environment 'authorAbstract'
The new definition of a 'abstract' environment in OLLbook.cls
has been removed.

Instead there is a new environment 'authorAbstract' in OLLstyles.sty.
This is used to attribute a chapter to its original author and provide an abstract.

The command \originalAuthor may be still used when it is needed
without the following abstract.

ulComplexPianoScores.tex has been updated accordingly