draftMode: New access technique
Implement a new technique for activating draft mode:
Each library file includes its corresponding draft mode file, if 'draft-mode' is defined in the master file.

TO BE DONE:
So far each library file has to
- supply the name of the draftMode file
- include draftMode.ily
It would be nice if draft-mode-file could be constructed from the
name of the original library file.