musicexample.sty: start MFE import
I want to loop through lilypond's systems in multi-system examples.

This commit provides a first solution to read in the system-count file that lilypond-book-preamble provides.

This can be used to set the counter in a for loop (to be created) that then will include the systems consecutively inside the example environment

TODO: 
- Implement the loop
- Check file name patterns