xmp: redesign \fullPageMusicExample
\fullPageMusicExample (and its ...Lilypond... counterpart) can now handle one- or multi-page pdf files.
A counter determines if we are on page 1 or any later page and sets \thispagestyle accordingly.

Also I remove the use of the scrplain page style for the subsequent pages and replace it with a custom style xmpPageRest, which seems more consistent.

The user has to ensure that these two page styles fit to the page styles of the rest of the document, though