xmp: Fix two-sided issue
Fixes #23
Now a full-page music example for which only a specific -odd or -even pdf is provided
reliably starts at the next odd/even page.
If necessary it launches a second \afterpage instance.

The problem was that I did the checks and \renewcommands within the first \afterpage block,
which obviously doesn't work.
So now I do all checks first and launch \afterpage then.