Make manual independent from oll-base
... by redundantly copying all needed commands in musicexamplesStyles.sty

Todo: Make manual independent from fontspec:
If compiled with other latexes than xetex or luatex simply omit the font loading.
We want the manual to be compiled with fontspec, but as there isn't a specific reason why this is necessary we want that anybody can compile it without our fonts.